+++ examples/miniapps/bundles/run.py
@@ -1,6 +1,6 @@
 """Example application - Bundles."""
 
-from core import containers
+from dependency_injector import containers
 from dependency_injector import providers
 
 from bundles.users import Users